scripts/requirements.txt
# TERRA Sneakers Importer - Dépendances Python
pandas>=2.0.0
requests>=2.28.0
Pillow>=9.5.0
python-dotenv>=1.0.0

# Optionnel pour datasets avancés
kaggle>=1.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
